Move component and test files
